requirements.txt
pydantic>=2.7.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
langchain>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
pinecone>=5.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0